backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-docx 
python-dotenv
pydantic
python-multipart>=0.0.9
anthropic>=0.30.0   # optional legacy fallback
elevenlabs
# speech support
faster-whisper>=0.8.0
transformers>=4.33.0
torch>=2.0.0
scipy>=1.10.0
langgraph>=0.10.0
psycopg2-binary
